# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Innoculus API specification — Manager / Editor / Verifier pipeline
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: jobs
    description: Job lifecycle and pipeline management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /jobs:
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a pipeline job
      description: Creates a new job and enqueues it for the Editor pipeline. Idempotent when job_id is supplied.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "200":
          description: Existing job returned (idempotent)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all jobs
      description: Returns a paginated list of all jobs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated job list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedJobsResponse"

  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job by ID
      description: Returns full job state including status, artifact reference, and diagnostics
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Job found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /jobs/{id}/status:
    patch:
      operationId: patchJobStatus
      tags: [jobs]
      summary: Update job status
      description: Editor or internal pipeline updates job pipeline status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchJobStatusRequest"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /jobs/{id}/artifact:
    put:
      operationId: putJobArtifact
      tags: [jobs]
      summary: Upload artifact
      description: Editor uploads a computed artifact. Creates a new immutable version.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutJobArtifactRequest"
      responses:
        "200":
          description: Artifact stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobArtifact"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /jobs/{id}/work:
    post:
      operationId: dispatchJobWork
      tags: [jobs]
      summary: Dispatch work to Editor
      description: Manager dispatches work payload to the Editor subagent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchWorkRequest"
      responses:
        "202":
          description: Work dispatched
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /jobs/{id}/verify:
    post:
      operationId: dispatchJobVerify
      tags: [jobs]
      summary: Dispatch artifact to Verifier
      description: Manager sends artifact and diagnostics to the Verifier subagent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchVerifyRequest"
      responses:
        "202":
          description: Verification dispatched
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /jobs/{id}/verdict:
    post:
      operationId: postJobVerdict
      tags: [jobs]
      summary: Receive Verifier verdict
      description: Verifier posts its verdict back to the Manager
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostVerdictRequest"
      responses:
        "200":
          description: Verdict recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error

    KernelParams:
      type: object
      properties:
        type:
          type: string
          enum: [gaussian, mellin]
        sigma:
          type: number
        alpha:
          type: number
      required:
        - type

    LatencyParams:
      type: object
      properties:
        lambda:
          type: number
        delta:
          type: number
        Tnow:
          type: number
      required: [lambda, delta, Tnow]

    TruncationParams:
      type: object
      properties:
        M:
          type: integer
        r:
          type: integer
      required: [M, r]

    PrecisionConfig:
      type: object
      properties:
        b:
          type: integer
        tol:
          type: number
        safety_margin:
          type: number
      required: [b, tol]

    PolicyConfig:
      type: object
      description: |
        Per-job policy thresholds. The numerical fields apply to the numerical
        Editor/Verifier; the cutoff_* fields apply to the cutoff_trace Verifier.
        Per-job values override defaults.
      properties:
        spectral_radius_max:
          type: number
          default: 0.999
        cond_limit:
          type: number
          default: 1000000
        dual_error_tol:
          type: number
          default: 0.000001
        spectral_tail_tol:
          type: number
          default: 0.000001
        judge_disagreement_max:
          type: number
          minimum: 0
          maximum: 1
          default: 0.34
          description: cutoff_trace CT02 — max allowed judge spot-recheck disagreement rate
        min_probes_per_month:
          type: integer
          minimum: 1
          default: 2
          description: cutoff_trace CT04 — minimum probes required per YYYY-MM bin
        min_recheck_count:
          type: integer
          minimum: 1
          default: 3
          description: cutoff_trace CT02 — minimum number of probes spot-rechecked
        warburg_residual_tol:
          type: number
          minimum: 0
          default: 0.05
          description: numerical CHK08 — max relative ||F − F̃||/||F̃|| against the closed-form Warburg oracle

    CreateJobRequest:
      oneOf:
        - $ref: "#/components/schemas/CreateNumericalJobRequest"
        - $ref: "#/components/schemas/CreateCutoffTraceJobRequest"
      discriminator:
        propertyName: kind

    CreateNumericalJobRequest:
      type: object
      properties:
        kind:
          type: string
          enum: [numerical]
          description: Optional kind discriminator. Defaults to numerical.
        job_id:
          type: string
          format: uuid
          description: Optional client-supplied ID for idempotency
        kernel:
          $ref: "#/components/schemas/KernelParams"
        Q:
          type: array
          items:
            type: array
            items:
              type: number
        truncation:
          $ref: "#/components/schemas/TruncationParams"
        latency:
          $ref: "#/components/schemas/LatencyParams"
        precision:
          $ref: "#/components/schemas/PrecisionConfig"
        policy_config:
          $ref: "#/components/schemas/PolicyConfig"
        model_pool:
          type: array
          items:
            type: object
            additionalProperties: true
      required: [kernel, Q, truncation, latency, precision]

    CutoffProbe:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
          description: Ground-truth answer for the LLM-as-judge to compare against
        date:
          type: string
          description: Real-world publication / event date in YYYY-MM-DD form
      required: [question, answer, date]

    CreateCutoffTraceJobRequest:
      type: object
      properties:
        kind:
          type: string
          enum: [cutoff_trace]
        job_id:
          type: string
          format: uuid
          description: Optional client-supplied ID for idempotency
        model:
          type: string
          description: Target model whose knowledge cutoff is being estimated
        judge_model:
          type: string
          description: Model used as the LLM-as-judge to grade target answers
        judge_temperature:
          type: number
          default: 0
        probes:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CutoffProbe"
        policy_config:
          $ref: "#/components/schemas/PolicyConfig"
      required: [kind, model, judge_model, probes]

    CutoffProbeResult:
      type: object
      properties:
        question: { type: string }
        answer: { type: string }
        date: { type: string }
        model_answer: { type: string }
        judge_score:
          type: number
          minimum: 0
          maximum: 1
        judge_reasoning: { type: string }
      required: [question, answer, date, model_answer, judge_score, judge_reasoning]

    MonthlyAggregate:
      type: object
      properties:
        month:
          type: string
          description: YYYY-MM
        n:
          type: integer
        knew_rate:
          type: number
          minimum: 0
          maximum: 1
      required: [month, n, knew_rate]

    CutoffEstimate:
      type: object
      properties:
        month: { type: string, description: YYYY-MM }
        ci_low: { type: string, description: YYYY-MM (95% CI lower bound) }
        ci_high: { type: string, description: YYYY-MM (95% CI upper bound) }
        fit_quality:
          type: number
          description: McFadden pseudo-R² of the logistic changepoint fit
      required: [month, ci_low, ci_high, fit_quality]

    NumericalArtifactPayload:
      type: object
      additionalProperties: true
      description: Numerical pipeline artifact (kernel coefficients, diagnostics, spectral data).

    CutoffArtifactPayload:
      type: object
      properties:
        kind:
          type: string
          enum: [cutoff_trace]
        model: { type: string }
        judge_model: { type: string }
        probe_results:
          type: array
          items:
            $ref: "#/components/schemas/CutoffProbeResult"
        monthly_aggregates:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyAggregate"
        cutoff_estimate:
          $ref: "#/components/schemas/CutoffEstimate"
      required: [kind, model, judge_model, probe_results, monthly_aggregates, cutoff_estimate]

    ArtifactPayload:
      oneOf:
        - $ref: "#/components/schemas/NumericalArtifactPayload"
        - $ref: "#/components/schemas/CutoffArtifactPayload"
      discriminator:
        propertyName: kind

    Job:
      type: object
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
          enum: [numerical, cutoff_trace]
          description: Job kind. numerical = original Editor pipeline; cutoff_trace = LLM knowledge-cutoff probing.
        status:
          type: string
          enum: [queued, editor_running, verifying, complete, complete_with_warnings, failed]
        kernel_params:
          type: object
          additionalProperties: true
          description: Job descriptor (numerical kernel params, or cutoff_trace probe specification).
        policy_config:
          type: object
          additionalProperties: true
        current_artifact_id:
          type: string
          format: uuid
          nullable: true
        retry_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, kind, status, kernel_params, policy_config, retry_count, created_at, updated_at]

    JobDetail:
      allOf:
        - $ref: "#/components/schemas/Job"
        - type: object
          properties:
            artifact:
              $ref: "#/components/schemas/JobArtifact"
            diagnostics:
              $ref: "#/components/schemas/JobDiagnostics"

    JobArtifact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        version:
          type: integer
        payload:
          $ref: "#/components/schemas/ArtifactPayload"
        hash:
          type: string
        signed_proof:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      required: [id, job_id, version, payload, hash, created_at]

    JobDiagnostics:
      type: object
      properties:
        id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        artifact_id:
          type: string
          format: uuid
        spectral_radius:
          type: number
        cond_i_minus_g:
          type: number
        dual_truncation_error:
          type: number
        spectral_tail_error:
          type: number
        verdict:
          type: string
          enum: [pass, warn, fail]
        issues:
          type: array
          items:
            $ref: "#/components/schemas/DiagnosticIssue"
        created_at:
          type: string
          format: date-time
      required: [id, job_id, artifact_id, spectral_radius, cond_i_minus_g, dual_truncation_error, spectral_tail_error, verdict, issues, created_at]

    DiagnosticIssue:
      type: object
      properties:
        check_id:
          type: string
        severity:
          type: string
          enum: [fail, warn]
        message:
          type: string
        remediation:
          type: string
      required: [check_id, severity, message, remediation]

    PaginatedJobsResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
      required: [jobs, total, page, page_size]

    PatchJobStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [queued, editor_running, verifying, complete, complete_with_warnings, failed]
        step:
          type: string
      required: [status]

    PutJobArtifactRequest:
      type: object
      properties:
        payload:
          type: object
          additionalProperties: true
        hash:
          type: string
      required: [payload, hash]

    DispatchWorkRequest:
      type: object
      properties:
        kernel:
          $ref: "#/components/schemas/KernelParams"
        Q:
          type: array
          items:
            type: array
            items:
              type: number
        truncation:
          $ref: "#/components/schemas/TruncationParams"
        seed:
          type: integer
      required: [kernel, Q, truncation]

    DispatchVerifyRequest:
      type: object
      properties:
        artifact_id:
          type: string
          format: uuid
        diagnostics:
          type: object
          additionalProperties: true
      required: [artifact_id]

    PostVerdictRequest:
      type: object
      properties:
        verdict:
          type: string
          enum: [pass, warn, fail]
        issues:
          type: array
          items:
            $ref: "#/components/schemas/DiagnosticIssue"
        recomputed_metrics:
          type: object
          additionalProperties: true
        signed_proof:
          type: string
      required: [verdict, issues, recomputed_metrics, signed_proof]